# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "XData Competitive & Clinical Intelligence Agent API",
  "description": "Evidence-grounded, continuously updated life-sciences competitive intelligence. Decision support only: facts are evidence-linked; interpretations are labelled AI hypotheses.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/entities/search": {
   "get": {
    "tags": [
     "entities"
    ],
    "summary": "Search",
    "operationId": "search_api_v1_entities_search_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 2,
       "title": "Q"
      }
     },
     {
      "name": "type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "company",
          "asset",
          "target",
          "mechanism",
          "indication",
          "biomarker",
          "endpoint",
          "modality",
          "line_of_therapy",
          "conference",
          "kol"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Type"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "default": 20,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response Search Api V1 Entities Search Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/entities/resolve": {
   "post": {
    "tags": [
     "entities"
    ],
    "summary": "Resolve",
    "operationId": "resolve_api_v1_entities_resolve_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ResolveIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Resolve Api V1 Entities Resolve Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/entities/review": {
   "get": {
    "tags": [
     "entities"
    ],
    "summary": "Review Queue",
    "operationId": "review_queue_api_v1_entities_review_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "pending_review",
       "title": "Status"
      }
     },
     {
      "name": "entity_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Type"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Review Queue Api V1 Entities Review Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/entities/review/{link_id}": {
   "post": {
    "tags": [
     "entities"
    ],
    "summary": "Decide",
    "description": "Approve/reject a machine-proposed match. Decisions are audited and reused by the resolver (FR-ENT-005).",
    "operationId": "decide_api_v1_entities_review__link_id__post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "link_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Link Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReviewDecision"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Decide Api V1 Entities Review  Link Id  Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/entities/{entity_type}/{entity_id}/aliases": {
   "post": {
    "tags": [
     "entities"
    ],
    "summary": "Add Entity Alias",
    "operationId": "add_entity_alias_api_v1_entities__entity_type___entity_id__aliases_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "entity_type",
      "in": "path",
      "required": true,
      "schema": {
       "enum": [
        "company",
        "asset",
        "target",
        "mechanism",
        "indication",
        "biomarker",
        "endpoint",
        "modality",
        "line_of_therapy",
        "conference",
        "kol"
       ],
       "type": "string",
       "title": "Entity Type"
      }
     },
     {
      "name": "entity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Entity Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AliasIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Add Entity Alias Api V1 Entities  Entity Type   Entity Id  Aliases Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/entities/aliases/{alias_id}": {
   "delete": {
    "tags": [
     "entities"
    ],
    "summary": "Remove Alias",
    "operationId": "remove_alias_api_v1_entities_aliases__alias_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "alias_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Alias Id"
      }
     },
     {
      "name": "reason",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Reason"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/entities/merge": {
   "post": {
    "tags": [
     "entities"
    ],
    "summary": "Merge",
    "operationId": "merge_api_v1_entities_merge_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MergeIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Merge Api V1 Entities Merge Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/entities/split": {
   "post": {
    "tags": [
     "entities"
    ],
    "summary": "Split",
    "operationId": "split_api_v1_entities_split_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SplitIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Split Api V1 Entities Split Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/entities/{entity_id}/graph": {
   "get": {
    "tags": [
     "entities"
    ],
    "summary": "Entity Graph",
    "operationId": "entity_graph_api_v1_entities__entity_id__graph_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "entity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Entity Id"
      }
     },
     {
      "name": "depth",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 3,
       "minimum": 1,
       "default": 1,
       "title": "Depth"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Entity Graph Api V1 Entities  Entity Id  Graph Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/entities/{entity_type}/{entity_id}": {
   "get": {
    "tags": [
     "entities"
    ],
    "summary": "Get Entity",
    "operationId": "get_entity_api_v1_entities__entity_type___entity_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "entity_type",
      "in": "path",
      "required": true,
      "schema": {
       "enum": [
        "company",
        "asset",
        "target",
        "mechanism",
        "indication",
        "biomarker",
        "endpoint",
        "modality",
        "line_of_therapy",
        "conference",
        "kol"
       ],
       "type": "string",
       "title": "Entity Type"
      }
     },
     {
      "name": "entity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Entity Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Entity Api V1 Entities  Entity Type   Entity Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/entities/kinds": {
   "get": {
    "tags": [
     "entities"
    ],
    "summary": "Kinds",
    "operationId": "kinds_api_v1_entities_kinds_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Kinds Api V1 Entities Kinds Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
   "AliasIn": {
    "properties": {
     "alias": {
      "type": "string",
      "title": "Alias"
     },
     "alias_type": {
      "type": "string",
      "enum": [
       "dev_code",
       "generic",
       "brand",
       "legacy",
       "abbreviation",
       "spelling",
       "synonym"
      ],
      "title": "Alias Type",
      "default": "synonym"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     }
    },
    "type": "object",
    "required": [
     "alias"
    ],
    "title": "AliasIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MergeIn": {
    "properties": {
     "entity_type": {
      "type": "string",
      "enum": [
       "company",
       "asset",
       "target",
       "mechanism",
       "indication",
       "biomarker",
       "endpoint",
       "modality",
       "line_of_therapy",
       "conference",
       "kol"
      ],
      "title": "Entity Type"
     },
     "source_id": {
      "type": "string",
      "format": "uuid",
      "title": "Source Id"
     },
     "target_id": {
      "type": "string",
      "format": "uuid",
      "title": "Target Id"
     },
     "reason": {
      "type": "string",
      "title": "Reason"
     }
    },
    "type": "object",
    "required": [
     "entity_type",
     "source_id",
     "target_id",
     "reason"
    ],
    "title": "MergeIn"
   },
...
   "ResolveIn": {
    "properties": {
     "mention": {
      "type": "string",
      "title": "Mention"
     },
     "entity_type": {
      "type": "string",
      "enum": [
       "company",
       "asset",
       "target",
       "mechanism",
       "indication",
       "biomarker",
       "endpoint",
       "modality",
       "line_of_therapy",
       "conference",
       "kol"
      ],
      "title": "Entity Type"
     }
    },
    "type": "object",
    "required": [
     "mention",
     "entity_type"
    ],
    "title": "ResolveIn"
   },
   "ReviewDecision": {
    "properties": {
     "decision": {
      "type": "string",
      "enum": [
       "approve",
       "reject"
      ],
      "title": "Decision"
     },
     "entity_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Entity Id",
      "description": "Target entity when approving (defaults to proposed)"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     },
     "scope": {
      "type": "string",
      "enum": [
       "tenant",
       "global"
      ],
      "title": "Scope",
      "default": "tenant"
     }
    },
    "type": "object",
    "required": [
     "decision"
    ],
    "title": "ReviewDecision"
   },
...
   "SplitIn": {
    "properties": {
     "entity_type": {
      "type": "string",
      "enum": [
       "company",
       "asset",
       "target",
       "mechanism",
       "indication",
       "biomarker",
       "endpoint",
       "modality",
       "line_of_therapy",
       "conference",
       "kol"
      ],
      "title": "Entity Type"
     },
     "entity_id": {
      "type": "string",
      "format": "uuid",
      "title": "Entity Id"
     },
     "new_name": {
      "type": "string",
      "title": "New Name"
     },
     "alias_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "minItems": 1,
      "title": "Alias Ids"
     },
     "reason": {
      "type": "string",
      "title": "Reason"
     }
    },
    "type": "object",
    "required": [
     "entity_type",
     "entity_id",
     "new_name",
     "alias_ids",
     "reason"
    ],
    "title": "SplitIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}